SubtitleTagging import: fixed bug where the last subtitle_mark spills into id_title:
This is an interesting bug. This behavior was caused by the diff-match-patch algorithm used by Suspension::TextReplayer.

When applying text changes from content_at to subtitle_tagging_import, the dmp algorithm considered the `G` of 'Greatest' a better fit than the `G` of 'Going', so it inserted the subtitle_mark in the wrong position.

DMP was confused because we're adding the entire id_title and id_paragraph section, and so some of the tokens can spill into that added text after replaying text changes.

To mitigate this, I just add two newlines to the end of subtitle_tagging_import text. This makes the endings of the two texts look more alike and helps DMP make the right decision. It doesn't matter that we're changing subtitle_tagging_import text because TextReplayer will overwrite the two added newlines anyways. They just help it do a better job.