Fix command `sync subtitle_mark_character_positions` so that it can be run stand alone on primary repo:
In commit 491b93a1c we introduced logic that would execute this command only on primary repo as part of `rt import` commands. Turns out I put that logic in the wrong place. I originally put it inside the `sync` command when it should have been in the calling `import` command. This commit moves the logic to the calling command and out of the `sync` command, so that the `sync` command can be run on any repo if it's not part of an `import` command.