Updated sync subtitles (temp 3):
This commit contains the next stage of subtitle sync code. It will also be replaced with updated code.

It is based on entire file contents instead of hunks from git diff.

It uses JaccardSimilarityComputer to measure string similarities for aligned subtitle pairs, and subtitle boundary alignment.

It also has an intermediate step between `align_subtitles` and `compute_operations` where it extracts groups of aligned subtitle pairs that form operations. The shortcoming here is that it was slow, didn’t do a good job aligning strings and computing asp boundaries, and it couldn’t handle operations that crossed paragraph boundaries.