Improved sync subtitles: SubtitleOperationsForFile::AlignSubtitlePairs:
When computing alignment type, be more picky when assigning :fully_aligned: Both left and right alignment have to be high, too. This helps prevent cases where a subtitle is marked as fully_aligned when it should really be left aligned because one word was added to the end from the next subtitle.