Added new Subtitle::Operation type: RecordIdChange:
We need this operation type to track subtitles that were moved to a different record. It is required to update record_ids in the STM CSV files. STM CSV files are based on information stored in the st-ops file. Record ids are included in all subtitle operations by default. However the record ids of any subtitles that just moved to a different record without any other content changes were not being updated in the STM CSV files. So we introduce a new operation type that adds record_id information to the st-ops file so that it can be transferred to STM CSV file for update.