PDF export custom kerning: Added our own RtNoLineBreak command:
If the latex \nolinebreak command is preceded by an \hspace, that space is removed, the penalty is issued and the space is reinserted. However, this only happens for positive values of hspace. For our custom kerning we have a number of negative hspaces, and those were just silently dropped. So we create our custom command which is identical to nolinebreak, except it works the same for negative and positive values.