Split::Subtitles: Commented out unused/hackish code:
I added a hack to the place where we transfer subtitles from the foreign aligned sentences to the foreign plain text. I did it to work around a condition I encountered when autosplitting a german file. 

Turns out there was a translation error. The translator inserted a period when it should have been a colon. Once I fixed the translation error, I didn’t need this code any more.

Commented out for now in case we may need it anyways.