Updated st autosplit: Handle headers with subtitles and headers in the middle of files:
Headers with subtitles: I updated the regex to allow for optional subtitle_marks before the header token.

Headers in the middle of files: Because of the way Suspension::Suspender works, it removes the newlines between a header and the preceding block level element so that both end up on the same line in filtered text. I have to adjust the plain text accordingly when that happens to avoid a Suspension text mismatch error.