st sync bug fix: When importing foreign subtitles, sync to correct primary st_sync_commit:
When we import foreign subtitles, we catch up on any st_ops that have been synced since the subtitle export.

The old (wrong) behavior used the most recent git commit in the primary repo as the foreign st_sync_commit. The new behavior uses the `to` git commit of the last st_ops file processed.