Fixed running of sync subtitles after an import subtitle on a foreign file:
After we import subtitles for a foreign file, we want to transfer any accumulated subtitle operations for that file. In the previous implementation, we applied st_ops before the post subtitle import validation. This broke the validation.

In the new implementation we run st sync on the foreign file after the post subtitle import validation.