Updated Process::Split::Subtitles: Don’t add subtitle marks into headers:
When transferring subtitles from foreign aligned sentences to plain text, change the order in which we capture text fragments: Capture spaces and newlines before partial matches. This will prevent adding of subtitles to previous line (e.g., a header).

Also replace a loop of sub!s with a single gsub!