Updated st sync policy around pending subtitle imports:
Currently we are faced with the issue that any foreign files with a subtitle export and pending subtitle import are excluded from st sync and don't get new st_ops applied to them until the pending subtitle import is complete. During this time the files are not allowed to be exported to table. We would like to eliminate that restriction, so that all foreign files can be exported to table at any time.

We change the existing policy in two ways:

* We apply any new st_ops to all foreign files, even those that have pending subtitle imports.
* As a consequence we can include all foreign files that participate in st_sync (st_sync_active is true) in every table export.

When we import subtitles for the file we do the following:

* There is a pre-import validation that makes sure that content AT and subtitle import plain text are identical. As before, very rarely there may be differences in plain text of content AT and subtitle import file. Those differences will be resolved manually (as we do currently).
* Import subtitles. This will update the subtitles, however it won't touch any content.
* Apply all accumulated st_ops to content AT. Flag any modified subtitles for review.
* Set exported_subtitles_at_st_sync_commit to null.
* Update st_sync_commit to the current one.