Updated RFile::ContentAt#subtitles: Split differently between Primary and foreign:
In primary files every paragraph begins with a subtitle, so when we extract a content AT file’s subtitles, we split on paragraph boundaries and subtitle marks. However for foreign files we can’t use that approach since not all foreign paragraphs start with subtitles.

This mismatch caused issues when extracting st-content_updates for table release where subtitles were considered identical as long as the text before the first double newline was identical.