Updated Validator::PdfExportConsistency: Fix handling of missing spaces because of Eagle dropcap:
With recent changes to how atx_header (“# header”) trailing newlines are handled, the old way of detecting a leading eagle broke. Since the newlines after headers may be different, the newline before the eagle may not be present. So I switched the #index method with a regex.

NOTE: The git diff does not display the eagle characters in the change.